+++ lib/tortoise/package/unsuback.ex
@@ -5,17 +5,15 @@ defmodule Tortoise.Package.Unsuback do
 
   alias Tortoise.Package
 
-  @type package_identifier :: 0x0001..0xFFFF
-
   @opaque t :: %__MODULE__{
             __META__: Package.Meta.t(),
-            identifier: package_identifier() | nil
+            identifier: Tortoise.package_identifier()
           }
   @enforce_keys [:identifier]
   defstruct __META__: %Package.Meta{opcode: @opcode, flags: 0b0000},
             identifier: nil
 
-  @spec decode(<<_::32>>) :: __MODULE__.t()
+  @spec decode(<<_::32>>) :: t
   def decode(<<@opcode::4, 0::4, 2, identifier::big-integer-size(16)>>)
       when identifier in 0x0001..0xFFFF do
     %__MODULE__{identifier: identifier}