+++ lib/tortoise/registry.ex
@@ -1,18 +1,17 @@
 defmodule Tortoise.Registry do
   @moduledoc false
 
-  @type client_id :: term()
-  @type via :: {:via, Registry, {__MODULE__, {module(), client_id()}}}
+  @type via :: {:via, Registry, {__MODULE__, {module(), Tortoise.client_id()}}}
   @type key :: atom() | tuple()
 
-  @spec via_name(module(), client_id()) :: via() | pid()
+  @spec via_name(module(), Tortoise.client_id()) :: via() | pid()
   def via_name(_module, pid) when is_pid(pid), do: pid
 
   def via_name(module, client_id) do
     {:via, Registry, reg_name(module, client_id)}
   end
 
-  @spec reg_name(module(), client_id()) :: {__MODULE__, {module(), client_id()}}
+  @spec reg_name(module(), Tortoise.client_id()) :: {__MODULE__, {module(), Tortoise.client_id()}}
   def reg_name(module, client_id) do
     {__MODULE__, {module, client_id}}
   end